requirements.txt
discord.py>=2.4.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
PyYAML>=6.0
better-profanity>=0.7.0
spotipy>=2.23.0
yt-dlp>=2024.1.0
rapidfuzz>=3.6.0
Pillow>=10.0.0
aiohttp>=3.9.0
